LayoutTests/accessibility/aria-required-updates-after-dynamic-change-expected.txt

This tests that changing the aria-required attribute is properly reflected in an AX object's required status.

#textbox is required: false
Setting aria-required to true on #textbox.
#textbox is required: true
Setting aria-required to false on #textbox.
#textbox is required: false

PASS successfullyParsed is true

TEST COMPLETE
